dév(kuri) : réduis la taille de InfoTypeMembreStructure
Ceci réduis la taille de InfoTypeMembreStructure de 8 octets par instance en changeant le type de "décalage" qui n'a pas besoin d'être sur 64-bit.
Le gain de 8 octets provient de la réduction du membre de 4 octets, et la suppression d'un rembourrage de 4 octets.

Ceci réduis d'environ 15% la taille utilisée pour les InfoTypeMembreStructures de Imago (19 376 octets -> 16 600 octets), et d'environ 15% également pour CréationBDD (186 536 octets -> 159 888 octets).

+++ logiciels/kuri/arbre_syntaxique/infos_types.hh
@@ -71,7 +71,7 @@ struct InfoTypeMembreStructure {
 
     kuri::chaine_statique nom{};
     InfoType *info = nullptr;
-    int64_t décalage = 0;  // décalage en octets dans la structure
+    int32_t décalage = 0;  // décalage en octets dans la structure
     int drapeaux = 0;
     kuri::tranche<const Annotation *> annotations{};
 };
